+++ webcentral/src/tools_over/translation.py
@@ -4,37 +4,9 @@ from modeltranslation.translator import (
 )
 
 from .models import (
-    Accessibility,
-    ApplicationArea,
-    Classification,
-    Focus,
-    LifeCyclePhase,
-    Scale,
-    TargetGroup,
     Tools,
-    Usage,
-    UserInterface,
 )
 
-
-class ApplicationAreaTranslationOptions(TranslationOptions):
-    fields = ("applicationArea",)
-
-
-class ClassificationTranslationOptions(TranslationOptions):
-    fields = ("classification",)
-
-
-class FocusTranslationOptions(TranslationOptions):
-    fields = ("focus",)
-
-
-class UsageTranslationOptions(TranslationOptions):
-    fields = ("usage",)
-
-
-class TargetGroupTranslationOptions(TranslationOptions):
-    fields = ("targetGroup",)
 
 
 class ToolsTranslationOptions(TranslationOptions):
@@ -50,29 +22,6 @@ class ToolsTranslationOptions(TranslationOptions):
     )
 
 
-class LifeCyclePhaseTranslationOptions(TranslationOptions):
-    fields = ("lifeCyclePhase",)
 
-
-class UserInterfaceTranslationOptions(TranslationOptions):
-    fields = ("userInterface",)
-
-
-class AccessibilityTranslationOptions(TranslationOptions):
-    fields = ("accessibility",)
-
-
-class ScaleTranslationOptions(TranslationOptions):
-    fields = ("scale",)
-
-
-translator.register(Accessibility, AccessibilityTranslationOptions)
-translator.register(ApplicationArea, ApplicationAreaTranslationOptions)
-translator.register(Classification, ClassificationTranslationOptions)
-translator.register(Focus, FocusTranslationOptions)
-translator.register(Usage, UsageTranslationOptions)
-translator.register(Scale, ScaleTranslationOptions)
-translator.register(TargetGroup, TargetGroupTranslationOptions)
 translator.register(Tools, ToolsTranslationOptions)
-translator.register(LifeCyclePhase, LifeCyclePhaseTranslationOptions)
-translator.register(UserInterface, UserInterfaceTranslationOptions)
+
